refactor(components/Link): migrate to LinkPrimitive
