+++ packages/components/src/components/Link/Link.test.tsx
@@ -4,7 +4,7 @@ import { screen, render } from '@testing-library/react';
 import { userEvent } from '@testing-library/user-event';
 import { describe, it, expect, vi } from 'vitest';
 
-import { Link } from './index.js';
+import { Link } from './Link';
 
 describe('Link', () => {
   const baseProps = {
