+++ packages/components/src/components/Link/Link.tsx
@@ -1,9 +1,9 @@
 'use client';
 
-import type { ComponentPropsWithRef, ComponentRef, ElementType } from 'react';
+import type { ComponentPropsWithRef, ElementType } from 'react';
 
-import { clsx, useDOMRef, polymorphicForwardRef } from '@koobiq/react-core';
-import { useLink } from '@koobiq/react-primitives';
+import { clsx, polymorphicForwardRef } from '@koobiq/react-core';
+import { Link as LinkPrimitive } from '@koobiq/react-primitives';
 
 import s from './Link.module.css';
 import type { LinkBaseProps } from './types';
@@ -13,48 +13,44 @@ export const Link = polymorphicForwardRef<'a', LinkBaseProps>((props, ref) => {
     variant = 'text-normal',
     visitable = false,
     pseudo = false,
+    disabled,
     as = 'a',
     startIcon,
     endIcon,
     children,
     className,
-    style,
+    ...other
   } = props;
-
-  const Tag = as;
-
-  const domRef = useDOMRef<ComponentRef<'a'>>(ref);
-
-  const elementType = as !== 'a' && as !== 'button' ? `${as}` : undefined;
-
-  const { linkProps, hovered, pressed, focusVisible } = useLink(
-    { ...props, elementType },
-    domRef
-  );
 
   const hasIcon = Boolean(startIcon || endIcon);
 
+  const elementType = as !== 'a' && as !== 'button' ? `${as}` : undefined;
+
   return (
-    <Tag
-      {...linkProps}
-      className={clsx(
-        s.base,
-        s[variant],
-        pseudo && s.pseudo,
-        hovered && s.hovered,
-        pressed && s.pressed,
-        hasIcon && s.hasIcon,
-        visitable && s.visitable,
-        focusVisible && s.focusVisible,
-        className
-      )}
-      style={style}
-      ref={domRef}
+    <LinkPrimitive
+      as={as}
+      disabled={disabled}
+      elementType={elementType}
+      className={({ hovered, pressed, focusVisible }) =>
+        clsx(
+          s.base,
+          s[variant],
+          pseudo && s.pseudo,
+          hovered && s.hovered,
+          pressed && s.pressed,
+          hasIcon && s.hasIcon,
+          visitable && s.visitable,
+          focusVisible && s.focusVisible,
+          className
+        )
+      }
+      {...other}
+      ref={ref}
     >
       {startIcon}
       {children}
       {endIcon}
-    </Tag>
+    </LinkPrimitive>
   );
 });
